+++ NEWS.md
@@ -1,12 +1,16 @@
 TimeStructures release notes
 ===================================
 
-Version 0.2.0 (2021-08-03)
+Version 0.2.0 (2021-08-04)
 --------------------------
 * Restructured and commented the file for improved readability (#7)
 * Inclusion of functions for DynamicTwoLevel and DynamicOperationalLevel (#7)
 * Inclusion of DynamicStrategicLevel (#7)
 * Export of all two level time structures (#7)
+* Bug fixes:
+    * Iterations over time structures and strategic periods result in 
+      proper duration for operational periods
+    * Functions next and previous now have proper durations
 
 Version 0.1.7 (2021-06-14)
 --------------------------